+++ src/extension.ts
@@ -1,6 +1,7 @@
 import * as vscode from 'vscode';
 import { contextServiceManager } from "./editor-context-service/manager"
 import * as listEditing from './listEditing';
+import * as formatting from './format';
 
 export function activate(context: vscode.ExtensionContext) {
 	context.subscriptions.push(
@@ -12,6 +13,9 @@ export function activate(context: vscode.ExtensionContext) {
     
     // Override `Enter`, `Tab` and `Backspace` keys
     listEditing.activate(context);
+
+    // Shortcuts
+    formatting.activate(context);
 
 }