+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Change Log
 
-All notable changes to Typst Companion will be documented in this file.
+All notable changes to Typst Companion will be documented here.
+
+## 0.0.3
+
+- Added toggle Bold with `ctrl/cmd + b`.
+- Added toggle Italic with `ctrl/cmd + i`.
+- Added toggle Underline with `ctrl/cmd + u`.
+- Added increase/decrease Header level with `ctrl/cmd + shift + ]` and `ctrl/cmd + shift + ]`, respectively.
+- Added 'Typst Companion: Toggle List' command to command palette.
+
+## 0.0.2
+
+- New logo.
 
 ## 0.0.1